Setelah pertanyaan

+++ daspro-jobsheet6/Pemilihan2Percobaan215.java
@@ -4,13 +4,13 @@ public class Pemilihan2Percobaan215 {
     public static void main(String[] args) {
         Scanner input15 = new Scanner(System.in);
         System.out.print("Masukkan sudut 1 : ");
-        int sudut1 = input15.nextInt();
+        float sudut1 = input15.nextFloat();
         System.out.print("Masukkan sudut 2 : ");
-        int sudut2 = input15.nextInt();
+        float sudut2 = input15.nextFloat();
         System.out.print("Masukkan sudut 3 : ");
-        int sudut3 = input15.nextInt();
+        float sudut3 = input15.nextFloat();
 
-        int totalSudut;
+        float totalSudut;
         totalSudut = sudut1 + sudut2 + sudut3;
 
         if (totalSudut == 180) {
